[modbus] exceptions moved to their own namespace
